Code clean up
